backend/requirements.txt
# FastAPI and Web
fastapi==0.121.0
uvicorn[standard]==0.38.0

# Database
sqlmodel==0.0.27
sqlalchemy==2.0.44

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
email-validator==2.3.0

# Configuration
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Forms
python-multipart==0.0.20

# NLP
yake==0.6.0

# HTTP
httpx==0.28.1

# Development & Testing
pytest==9.0.1
pytest-cov==7.0.0
pytest-asyncio==1.3.0
mypy==1.18.2
ruff==0.14.5
types-python-jose==3.3.4.20240106
types-passlib==1.7.7.20240106
